Fix destroy method breaking rebuild
Problem: We don't set the `since` value during `destroy()`, which means
that the rebuild method from FlumeDB doesn't actually pass us any of the
old messages. Flumeview-Level just deletes what's in the database and
then never sees the old messages again.

Solution: Use the new `clear()` method from the new version of LevelDB
instead of deleting the database, reducing complexity, and fix the
`destroy()` method so that it can actually do a rebuild.